# api.yml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'testManagerPI'
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: 'localhost:3000'
paths:
...
    /player/team:
        post:
            summary: 'Add Player to a Team'
            operationId: SetTeam
            requestBody:
                content:
                    'application/json':
                        schema:
                            $ref: "#/components/schemas/teamInfo"
                        examples:
                             teamInfo1:
                                        summary: test_id, player_id pair
                                        value: {"player_id": 2}

            responses:
                '200':
                    description: 'Sample response: Manage a team'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Player'
                            examples:
                                testPlayer:
                                    summary: team members details
                                    value: {"player_id": 1,
                                                "firstname": "Louis",
                                                "surname": "Dem",
                                                "age": 22,
                                                "education": "Admin"}
                default:
                    description: Unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        delete:
            summary: 'Remove Player from a Team'
            operationId: DeleteFromTeam
            requestBody:
                content:
                    'application/json':
                        schema:
                            $ref: "#/components/schemas/teamInfo"
                        examples:
                             teamInfo1:
                                        summary: test_id, player_id pair
                                        value: {"player_id": 2}
            responses:
                '200':
                    description: 'Sample response: Manage a team'
                    content:
                        application/json:
                            schema:
                                type: string
                default:
                    description: Unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    Player:
      type: object
      required:
        - player_id
        - firstname
        - surname
        - age
        - education
      properties:
        player_id:
          type: integer
        firstname:
            type: string
        surname:
            type: string
        age:
            type: integer
        education:
            type: string
    teamInfo:
        type : object
        required:
            - test_id
            - player_id
        properties:
            test_id:
                type: integer
            player_id:
                type: integer
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string